Update Dynamo For Revit to support 64-bit ElementId. (#2849)
* Update Dynamo For Revit to support 64-bit ElementId.

- Address all Deprecation warnings regarding use of old IntegerValue property.
- Address all Deprecation warnings regarding construction of ElementId using Int.
- Update Dynamo wrappers to treat Id's as long (64-bit) not int.

* Fix regression failure

+++ test/Libraries/RevitNodesTests/Elements/FloorTypeTests.cs
@@ -73,7 +73,7 @@ namespace RevitNodesTests.Elements
             var floorType = floor.ElementType as FloorType;
             var foundationType = foundationSlab.ElementType as FloorType;
             var structuralMaterial = floorType.GetStructuralMaterial();
-            int structuralMaterailId = structuralMaterial.Id;
+            long structuralMaterailId = structuralMaterial.Id;
             var foundationSlabException = Assert.Throws<System.InvalidOperationException>(() => foundationType.GetStructuralMaterial());
 
             // Assert
